Move convolution functions to dedicated file
This will make it slightly simpler to build the convolution functions
into host programs for testing.
