+++ hosttest/Makefile
@@ -3,6 +3,8 @@ all: streamtest
 TARGET = ..
 CFLAGS = -O3 -g3 -I../include/
 
+LIBOBJS := $(TARGET)/convolution.o $(TARGET)/symbol.o $(TARGET)/sample.o
+
 %.o: %.c
 	gcc -c $(CFLAGS) -o $@ $^
 
@@ -10,10 +12,10 @@ CFLAGS = -O3 -g3 -I../include/
 %.pico: %.c
 	gcc -c $(CFLAGS) -fpic -o $@ $^
 
-streamtest: streamtest.o $(TARGET)/symbol.o $(TARGET)/sample.o
+streamtest: streamtest.o $(LIBOBJS)
 	gcc $(CFLAGS) -o $@ $^
 
-streamtest.so: streamtest.pico $(TARGET)/symbol.pico $(TARGET)/sample.pico
+streamtest.so: streamtest.pico $(LIBOBJS:.o=.pico)
 	gcc -shared -o $@ $^
 
 plotpng: plotdata
