+++ include/sample.h
@@ -27,9 +27,7 @@ struct filter_stream {
 
 #define to_filter_stream(__ss) container_of(__ss, struct filter_stream, stream)
 
-float convolve_point(float *in, float *fir, int in_len, int out_idx, int fir_len);
 int fill_iq_bufs(struct symbol_stream *ss, float *i, float *q, int n);
-void convolve(float *in, float *out, float *fir, int in_len, int out_len, int fir_len);
 int init_filter_stream(struct filter_stream *fs, struct symbol_stream *ss, float *coeffs, int filter_len);
 int init_sample_hold(struct sample_hold *sh, struct symbol_stream *ss, int hold_count);
